Made a sidenav in the index and organized css
In the index, a nav bar has been made on the left of the page with links to pages that haven't yet been made. In the gitPractice.css file, the code has been separated into the main bits used for every page and the bits used only for the login and sign up pages.