Created the everythingGit page
The everythingGit page was created, as well as some images saved which were added to the page. A new section was added to the CSS file to accomodate this. The margin on the bottom of the sidenav was also lengthened in order to prevent the text falling to the left side of the page.